refactor loading spinner
