+++ src/auth/Authentication.jsx
@@ -1,14 +1,6 @@
 import React, { useEffect, useState } from "react";
-import { Container, Spinner } from "react-bootstrap";
+import LoadingSpinner from '../components/LoadingSpinner'
 import firebase from "../firebaseConfig";
-
-const spinnerStyle = {
-  height: "100vh",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  color: "#fe8f7c",
-};
 
 export const AuthContext = React.createContext();
 
@@ -23,17 +15,9 @@ export const AuthProvider = ({ children }) => {
     });
   }, []);
 
-  if (pending) {
-    return (
-      <Container fluid style={spinnerStyle}>
-        <Spinner animation="border" role="status" size="md">
-          <span className="sr-only">Loading...</span>
-        </Spinner>
-      </Container>
-    );
-  }
-
   return (
-    <AuthContext.Provider value={currentUser}>{children}</AuthContext.Provider>
+    pending
+    ? <LoadingSpinner />
+    : <AuthContext.Provider value={currentUser}>{children}</AuthContext.Provider>
   );
 };
